tests/i915/i915_pm_rc6_residency: Use correct type for power in rc6-fence
Use the correct type in rc6-fence for power to prevent truncating
value to 0.

+++ tests/i915/i915_pm_rc6_residency.c
@@ -509,7 +509,7 @@ static void rc6_fence(int i915, const intel_ctx_t *ctx)
 		igt_debug("%s: slept=%lu perf=%"PRIu64", rc6=%"PRIu64"\n",
 			  e->name, slept, ts[1] - ts[0], rc6);
 		if (rapl_read(&rapl, &sample[1]))  {
-			uint64_t power = power_J(&rapl, &sample[0], &sample[1]);
+			double power = power_J(&rapl, &sample[0], &sample[1]);
 			igt_info("Total energy used for %s: %.1fmJ (%.1fmW)\n",
 				 e->name,
 				 power * 1e3,
